collisions, explosions, and asteroid splitting
getting kinda close to finished :O
it's basically asteroids now.. but without levels or lives or ufos.

this commit took a bit of trial and error to implement, but I'm happy with
how it turned out. the explosions are a placeholder that I want to add
some more animation to in the future, but they look pretty cool as is