a few small changes and updates
worth noting that I finally actually once and for all fixed that bug
with disappearing/reappearing asteroids which I thought was related to
pointers. ...It turned out to be because I was using raylib's
Vector2Invert instead of Vector2Negate.

Using a debugger, it showed the bugged asteroid always had an abnormally huge
position vector, which meant it had to do with vector positioning rather
than pointers and memory allocation.